+++ src/lib/djerba/plugins/supplement/body/plugin.py
@@ -21,7 +21,10 @@ class main(plugin_base):
     GENETICIST = "clinical_geneticist_name"
     GENETICIST_ID = "clinical_geneticist_licence"
     EXTRACT_DATE = "extract_date"
-    
+
+    GENETICIST_DEFAULT = 'Trevor Pugh, PhD, FACMG'
+    GENETICIST_ID_DEFAULT = '1027812'
+
     def check_assay_name(self, wrapper):
         [ok, msg] = assays.name_status(wrapper.get_my_string(self.ASSAY))
         if not ok:
@@ -39,7 +42,10 @@ class main(plugin_base):
         if wrapper.my_param_is_null(self.ASSAY):
             if input_data:
                 wrapper.set_my_param(self.ASSAY, input_data[self.ASSAY])
-            else:
+            elif self.workspace.has_file("sample_info.json"):
+                sample_info = self.workspace.read_json("sample_info.json")
+                wrapper.set_my_param(self.ASSAY, sample_info[self.ASSAY])
+            else: 
                 msg = "Cannot find assay from input_params.json or manual config"
                 self.logger.error(msg)
                 raise DjerbaPluginError(msg)
@@ -91,8 +97,8 @@ class main(plugin_base):
             self.add_ini_discovered(key)
         self.set_ini_default(self.REPORT_SIGNOFF_DATE, self.NONE_SPECIFIED)
         self.set_ini_default(self.USER_SUPPLIED_DRAFT_DATE, self.NONE_SPECIFIED)
-        self.set_ini_default(self.GENETICIST, 'Trevor Pugh, PhD, FACMG')
-        self.set_ini_default(self.GENETICIST_ID, '1027812')
+        self.set_ini_default(self.GENETICIST, self.GENETICIST_DEFAULT)
+        self.set_ini_default(self.GENETICIST_ID, self.GENETICIST_ID_DEFAULT)
         self.set_ini_default(core_constants.ATTRIBUTES, 'clinical')
         self.set_ini_default(self.FAILED, "False")
         self.set_priority_defaults(self.DEFAULT_CONFIG_PRIORITY)