+++ src/lib/djerba/plugins/wgts/snv_indel/constants.py
@@ -4,7 +4,9 @@ TUMOUR_ID = 'tumour_id'
 NORMAL_ID = 'normal_id'
 MAF_PATH = 'maf_path'
 CNA_PATH = 'cna_path'
+HAS_COPY_STATE_DATA = 'has copy state data'
 HAS_EXPRESSION_DATA = 'has expression data'
+DEFAULT = 'snv_indel_default'
 
 # important MAF headers
 VARIANT_CLASSIFICATION = 'Variant_Classification'