+++ src/lib/djerba/plugins/wgts/snv_indel/plugin.py
@@ -27,33 +27,40 @@ class main(plugin_base):
     def configure(self, config):
         config = self.apply_defaults(config)
         wrapper = self.get_config_wrapper(config)
-        wrapper = self.update_wrapper_if_null(
-            wrapper,
-            input_params_helper.INPUT_PARAMS_FILE,
-            sic.ONCOTREE_CODE,
-            input_params_helper.ONCOTREE_CODE
-        )
-        wrapper = self.update_wrapper_if_null(
-            wrapper,
-            input_params_helper.INPUT_PARAMS_FILE,
-            sic.STUDY_ID,
-            input_params_helper.STUDY
-        )
+        # required params -- must be in INI or JSON
+        # MAF input file, required for obvious reasons
         wrapper = self.update_wrapper_if_null(
             wrapper,
             core_constants.DEFAULT_PATH_INFO,
             sic.MAF_PATH,
             'variantEffectPredictor_matched'
         )
+        # oncotree code is required for making OncoKB links and annotation
+        wrapper = self.update_wrapper_if_null(
+            wrapper,
+            input_params_helper.INPUT_PARAMS_FILE,
+            sic.ONCOTREE_CODE,
+            input_params_helper.ONCOTREE_CODE
+        )
+        # tumour ID is required for MAF update and OncoKB annotation
         wrapper = self.update_wrapper_if_null(
             wrapper,
             core_constants.DEFAULT_SAMPLE_INFO,
             sic.TUMOUR_ID
         )
+        # optional params with fallback value -- used only for constructing Whizbam links
+        wrapper = self.update_wrapper_if_null(
+            wrapper,
+            input_params_helper.INPUT_PARAMS_FILE,
+            sic.STUDY_ID,
+            input_params_helper.STUDY,
+            fallback=sic.DEFAULT
+        )
         wrapper = self.update_wrapper_if_null(
             wrapper,
             core_constants.DEFAULT_SAMPLE_INFO,
-            sic.NORMAL_ID
+            sic.NORMAL_ID,
+            fallback=sic.DEFAULT
         )
         if wrapper.my_param_is_null(sic.WHIZBAM_PROJECT):
             # if whizbam project not manually configured, default to study id
@@ -72,7 +79,6 @@ class main(plugin_base):
         # - Construct data for the gene info and treatment option mergers
         # - Make the VAF plot and record as base64
         wrapper = self.get_config_wrapper(config)  
-        work_dir = self.workspace.get_work_dir()
         data = self.get_starting_plugin_data(wrapper, self.PLUGIN_VERSION)
         whizbam_url = whizbam.link_base(
             sic.WHIZBAM_BASE_URL,
@@ -82,7 +88,7 @@ class main(plugin_base):
             self.SEQTYPE,
             self.GENOME
         )
-        proc = snv_indel_processor(work_dir, wrapper, self.log_level, self.log_path)
+        proc = snv_indel_processor(self.workspace, wrapper, self.log_level, self.log_path)
         proc.write_working_files(whizbam_url)
         data['results'] = proc.get_results()
         data['merge_inputs'] = proc.get_merge_inputs()